Change endpoint analyse to data
