+++ Uberback/Endpoint/Data.cs
@@ -1,104 +1,181 @@
-﻿using Google.Cloud.Vision.V1;
-using Nancy;
+﻿using Nancy;
 using Newtonsoft.Json;
+using RethinkDb.Driver.Net;
 using System;
 using System.Collections.Generic;
+using System.Collections.Specialized;
+using System.Globalization;
 using System.Linq;
-using System.Net.Http;
-using System.Text;
+using System.Threading.Tasks;
 
 namespace Uberback.Endpoint
 {
     public class Data : NancyModule
     {
-        /// <summary>
-        /// Post informationx
-        /// </summary>
+        public class AnalyseBatchRequestData
+        {
+            public string Content { get; set; }
+            public int Nb { get; set; }
+        }
+
+        public class AnalyseBatchRequestUrlBatch
+        {
+            public string UrlSrc { get; set; }
+            public List<AnalyseBatchRequestData> Images { get; set; }
+            public List<AnalyseBatchRequestData> Texts { get; set; }
+        }
+
+        public class AnalyseBatchRequest
+        {
+            public string Token { get; set; }
+            public string UserId { get; set; }
+            public string Service { get; set; }
+            public List<AnalyseBatchRequestUrlBatch> UrlBatchs { get; set; }
+        }
+
         public Data() : base("/data")
         {
-            Post("/", x =>
+            base.Post("/", async x =>
             {
                 var args = Common.ParseArgs(Request.Body);
 
-                // Error Handling
-                Common.Answer? error = Common.BasicCheck(args.Get("token"));
-                if (error.HasValue)
-                    return (Response.AsJson(new Response.Error()
-                    {
-                        Message = error.Value.message
-                    }, error.Value.code));
-                if (string.IsNullOrEmpty(args.Get("type")) || string.IsNullOrEmpty(args.Get("userId")) || string.IsNullOrEmpty(args.Get("content")) || string.IsNullOrEmpty(args.Get("token")) || string.IsNullOrEmpty(args.Get("service")))
-                    return (Response.AsJson(new Response.Error()
-                    {
-                        Message = "Missing arguments"
-                    }, HttpStatusCode.BadRequest));
-
-                List<string> flags = new List<string>();
-                // Add information to the db
-                switch (args.Get("type").ToString())
+                AnalyseBatchRequest batch;
+                try
                 {
-                    case "text":
-                        string finalMsg = Program.P.translationClient.TranslateText(args.Get("content"), "en").TranslatedText;
-                        using (HttpClient hc = new HttpClient())
-                        {
-                            HttpResponseMessage post = hc.PostAsync("https://commentanalyzer.googleapis.com/v1alpha1/comments:analyze?key=" + Program.P.perspectiveApi, new StringContent(
-                                    JsonConvert.DeserializeObject("{comment: {text: \"" + EscapeString(finalMsg) + "\"},"
-                                                                + "languages: [\"en\"],"
-                                                                + "requestedAttributes: {" + string.Join(":{}, ", categories.Select(y => y.Item1)) + ":{}} }").ToString(), Encoding.UTF8, "application/json")).GetAwaiter().GetResult();
+                    batch = ParseBatchArgs(args);
 
-                            dynamic json = JsonConvert.DeserializeObject(post.Content.ReadAsStringAsync().GetAwaiter().GetResult());
-                            foreach (var s in categories)
-                            {
-                                double value = json.attributeScores[s.Item1].summaryScore.value;
-                                if (value >= s.Item2)
-                                    flags.Add(s.Item1);
-                            }
-                            if (flags.Count == 0)
-                                flags.Add("SAFE");
-                        }
-                        Program.P.db.AddTextAsync(string.Join(",", flags), args.Get("userId"), args.Get("service")).GetAwaiter().GetResult();
-                        break;
-
-                    case "image":
-                        var image = Image.FetchFromUri(args.Get("content"));
-                        SafeSearchAnnotation response = Program.P.imageClient.DetectSafeSearch(image);
-                        if (response.Adult > Likelihood.Possible)
-                            flags.Add("Adult");
-                        if (response.Medical > Likelihood.Possible)
-                            flags.Add("Medical");
-                        if (response.Racy > Likelihood.Possible)
-                            flags.Add("Racy");
-                        if (response.Violence > Likelihood.Possible)
-                            flags.Add("Violence");
-                        if (flags.Count == 0)
-                            flags.Add("SAFE");
-                        Program.P.db.AddImageAsync(string.Join(",", flags), args.Get("userId"), args.Get("service")).GetAwaiter().GetResult();
-                        break;
-
-                    default:
-                        return (Response.AsJson(new Response.Error()
-                        {
-                            Message = "Type must be text or image"
-                        }, HttpStatusCode.BadRequest));
+                    // Check request
+                    Validator.Data.ValidatorResponse error = null;
+                    if ((error = Validator.Data.ValidateRequest(batch)) != null)
+                        return Response.AsJson(new Response.Error() { Message = error.Message }, error.StatusCode);
                 }
-                return (Response.AsJson(new Response.Empty(), HttpStatusCode.NoContent));
+                catch (Exception e)
+                {
+                    return Response.AsJson(new Response.Error() { Message = e.Message }, HttpStatusCode.BadRequest);
+                }
+
+                // Do request
+                var taskErrors = new List<Task<string>>();
+                foreach (var urlBatch in batch.UrlBatchs) {
+                    if (urlBatch.Images != null) {
+                        foreach (var image in urlBatch.Images) {
+                            taskErrors.Add(ConnectToAPIForAnalyseImageAsync(batch.UserId, urlBatch.UrlSrc, image.Content, batch.Service));
+                        }
+                    }
+                    if (urlBatch.Texts != null) {
+                        foreach (var text in urlBatch.Texts) {
+                            taskErrors.Add(ConnectToAPIForAnalyseTextAsync(batch.UserId, urlBatch.UrlSrc, text.Content, batch.Service));
+                        }
+                    }
+                }
+                await Task.WhenAll(taskErrors);
+                var errorsList = new List<string>();
+                foreach (var response in taskErrors) {
+                    if (response.Result != null) {
+                        errorsList.Add(response.Result);
+                    }
+                }
+                if (errorsList.Count != 0)
+                    return Response.AsJson(new Response.ErrorArray() { Message = errorsList.ToArray() }, HttpStatusCode.BadRequest);
+                else
+                    return Response.AsJson(new Response.Empty(), HttpStatusCode.NoContent);
             });
         }
 
-        private static string EscapeString(string msg)
+        public AnalyseBatchRequest ParseBatchArgs(NameValueCollection args)
         {
-            return msg.Replace("\\", "\\\\").Replace("\"", "\\\"");
+            var batch = new AnalyseBatchRequest
+            {
+                UrlBatchs = new List<AnalyseBatchRequestUrlBatch>()
+            };
+
+            batch.Token = args.Get("token");
+            batch.UserId = args.Get("userId");
+            batch.Service = args.Get("service");
+
+            int idxUrlBatch = 0;
+            string urlSrc;
+            while (!string.IsNullOrEmpty((urlSrc = args.Get("urlBatchs[" + idxUrlBatch + "][urlSrc]"))))
+            {
+                batch.UrlBatchs.Add(new AnalyseBatchRequestUrlBatch { UrlSrc = urlSrc, Images = new List<AnalyseBatchRequestData>(), Texts = new List<AnalyseBatchRequestData>() });
+
+                int idxData = 0;
+                while (!string.IsNullOrEmpty(args.Get("urlBatchs[" + idxUrlBatch + "][texts][" + idxData + "][nb]")))
+                {
+                    batch.UrlBatchs[idxUrlBatch].Texts.Add(new AnalyseBatchRequestData
+                    {
+                        Content = args.Get("urlBatchs[" + idxUrlBatch + "][texts][" + idxData + "][content]"),
+                        Nb = Int16.Parse(args.Get("urlBatchs[" + idxUrlBatch + "][texts][" + idxData + "][nb]"))
+                    });
+                    ++idxData;
+                }
+
+                idxData = 0;
+                while (!string.IsNullOrEmpty(args.Get("urlBatchs[" + idxUrlBatch + "][images][" + idxData + "][nb]")))
+                {
+                    batch.UrlBatchs[idxUrlBatch].Images.Add(new AnalyseBatchRequestData
+                    {
+                        Content = args.Get("urlBatchs[" + idxUrlBatch + "][images][" + idxData + "][content]"),
+                        Nb = Int16.Parse(args.Get("urlBatchs[" + idxUrlBatch + "][images][" + idxData + "][nb]"))
+                    });
+                    ++idxData;
+                }
+                ++idxUrlBatch;
+            }
+            return batch;
         }
 
-        private readonly Tuple<string, float>[] categories = new Tuple<string, float>[] {
-            new Tuple<string, float>("TOXICITY", .80f),
-            new Tuple<string, float>("SEVERE_TOXICITY", .60f),
-            new Tuple<string, float>("IDENTITY_ATTACK", .60f),
-            new Tuple<string, float>("INSULT", .60f),
-            new Tuple<string, float>("PROFANITY", .80f),
-            new Tuple<string, float>("THREAT", .60f),
-            new Tuple<string, float>("INFLAMMATORY", .60f),
-            new Tuple<string, float>("OBSCENE", .9f)
-        };
+        public static async Task<string> ConnectToAPIForAnalyseImageAsync(string userId, string urlSrc, string data, string service)
+        {
+            var imageUrl = Common.IsAbsoluteUrl(data) ? data : urlSrc + data;
+            try
+            {
+                var trigeredFlags = await Program.P.ImageAnalyser.AnalyseImageUrlAsync(imageUrl);
+                var flags = StringifyFlags(trigeredFlags);
+
+                Program.P.db.AddImageAsync(flags, userId, service).GetAwaiter().GetResult();
+                return null;
+            }
+            catch (Exception e)
+            {
+                return "Error in the analyze of image \"" + imageUrl + "\": " + e.Message;
+            }
+        }
+
+        public static async Task<string> ConnectToAPIForAnalyseTextAsync(string userId, string urlSrc, string text, string service)
+        {
+            try
+            {
+                string flags;
+                var textWithSalt = text + "0nes@l7yb0y¨^";
+                string hashedText = Common.GetHashString(textWithSalt);
+                // Check if the text is already analysed
+                if (await Program.P.db.IsTextAnalysedAsync(hashedText))
+                {
+                    // Get old Flag
+                    flags = await Program.P.db.GetFlagsFromAnalysedTextAsync(hashedText);
+                    await Program.P.db.UpdateLastDateTimeOfAnalysedTextAsync(hashedText);
+                }
+                else
+                {
+                    // Analyse text
+                    var trigeredFlags = await Program.P.TextAnalyser.AnalyseTextAsync(text);
+                    flags = StringifyFlags(trigeredFlags);
+
+                    // Store the result
+                    await Program.P.db.AddAnalysedTextAsync(flags, hashedText);
+                }
+                // Log the analysed text
+                await Program.P.db.AddTextAsync(flags, userId, service);
+                return null;
+            }
+            catch (Exception e)
+            {
+                return e.Message;
+            }
+        }
+
+        private static string StringifyFlags(Dictionary<string, string> trigeredFlags)
+            => string.Join(",", trigeredFlags.Keys);
     }
 }
