+++ Uberback/Validator/Data.cs
@@ -6,7 +6,7 @@ using System.Collections.Specialized;
 
 namespace Uberback.Validator
 {
-    class AnalyseBatch
+    class Data
     {
         public class ValidatorResponse
         {
@@ -14,11 +14,11 @@ namespace Uberback.Validator
             public HttpStatusCode StatusCode { get; set; }
         }
 
-        public static ValidatorResponse ValidateRequest(Endpoint.AnalyzeBatch.AnalyseBatchRequest args)
+        public static ValidatorResponse ValidateRequest(Endpoint.Data.AnalyseBatchRequest args)
         {
             ValidatorResponse error;
             if ((error = ValidateBatchRequest(args)) != null
-                ||(error = ValidateToken(args)) != null
+                || (error = ValidateToken(args)) != null
                 || (error = ValidateUserId(args)) != null
                 || (error = ValidateService(args)) != null
                 || (error = ValidateUrlBatch(args)) != null)
@@ -26,14 +26,14 @@ namespace Uberback.Validator
             return null;
         }
 
-        public static ValidatorResponse ValidateBatchRequest(Endpoint.AnalyzeBatch.AnalyseBatchRequest args)
+        public static ValidatorResponse ValidateBatchRequest(Endpoint.Data.AnalyseBatchRequest args)
         {
             if (args == null)
                 return new ValidatorResponse { Message = "Invalid format of request", StatusCode = HttpStatusCode.BadRequest };
             return null;
         }
 
-        public static ValidatorResponse ValidateUrlBatch(Endpoint.AnalyzeBatch.AnalyseBatchRequest args)
+        public static ValidatorResponse ValidateUrlBatch(Endpoint.Data.AnalyseBatchRequest args)
         {
             if (args.UrlBatchs == null || args.UrlBatchs.Count == 0)
                 return new ValidatorResponse { Message = "No Batch provided", StatusCode = HttpStatusCode.BadRequest };
@@ -50,31 +50,31 @@ namespace Uberback.Validator
             return null;
         }
 
-        public static ValidatorResponse ValidateImages(Endpoint.AnalyzeBatch.AnalyseBatchRequestUrlBatch args, string urlSrc)
+        public static ValidatorResponse ValidateImages(Endpoint.Data.AnalyseBatchRequestUrlBatch args, string urlSrc)
         {
             if (args.Images == null)
                 return null;
             foreach (var item in args.Images)
             {
-                if (item.Nb == 0 || string.IsNullOrEmpty(item.Data))
+                if (item.Nb == 0 || string.IsNullOrEmpty(item.Content))
                     return new ValidatorResponse { Message = "Invalid item in the list of images of \"" + urlSrc + "\": doesn't contain \"data\" or \"nb\"", StatusCode = HttpStatusCode.BadRequest };
             }
             return null;
         }
 
-        public static ValidatorResponse ValidateTexts(Endpoint.AnalyzeBatch.AnalyseBatchRequestUrlBatch args, string urlSrc)
+        public static ValidatorResponse ValidateTexts(Endpoint.Data.AnalyseBatchRequestUrlBatch args, string urlSrc)
         {
             if (args.Texts ==  null)
                 return null;
             foreach (var item in args.Texts)
             {
-                if (item.Nb == 0 || string.IsNullOrEmpty(item.Data))
+                if (item.Nb == 0 || string.IsNullOrEmpty(item.Content))
                     return new ValidatorResponse { Message = "Invalid item in the list of texts of \"" + urlSrc + "\": doesn't contain \"data\" or \"nb\"", StatusCode = HttpStatusCode.BadRequest };
             }
             return null;
         }
 
-        public static ValidatorResponse ValidateToken(Endpoint.AnalyzeBatch.AnalyseBatchRequest args)
+        public static ValidatorResponse ValidateToken(Endpoint.Data.AnalyseBatchRequest args)
         {
             Endpoint.Common.Answer? error = Endpoint.Common.BasicCheck(args.Token);
             if (error.HasValue)
@@ -82,7 +82,7 @@ namespace Uberback.Validator
             return null;
         }
 
-        public static ValidatorResponse ValidateUserId(Endpoint.AnalyzeBatch.AnalyseBatchRequest args)
+        public static ValidatorResponse ValidateUserId(Endpoint.Data.AnalyseBatchRequest args)
         {
             if (string.IsNullOrEmpty(args.UserId))
                 return new ValidatorResponse { Message = "The userId is not provided", StatusCode = HttpStatusCode.BadRequest };
@@ -91,7 +91,7 @@ namespace Uberback.Validator
             return null;
         }
 
-        public static ValidatorResponse ValidateService(Endpoint.AnalyzeBatch.AnalyseBatchRequest args)
+        public static ValidatorResponse ValidateService(Endpoint.Data.AnalyseBatchRequest args)
         {
             if (string.IsNullOrEmpty(args.Service))
                 return new ValidatorResponse { Message = "The Service is not provided", StatusCode = HttpStatusCode.BadRequest };
